patch: fixed imports skip
